+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Changes
 * [UI]: Sample modified date is now updated when sample metadata is modified.
 * [Developer]: Fix dependabot warning for `set-value`, force using a set-value above 4.0.1
 * [Developer]: Updated chromedriver to v93.0.1.
+* [Database]: Updated query to improve performance for metadata in REST API and line list.
 
 21.01 to 21.05
 --------------